Added Gemlock & migrations. Fleshed out pollTurk.
Added Gemlock file as I get closer to deploying on Heroku
Added file for migrations, but no rakefile

Changed how pollTurk is to work.
Using pingdom, I can GET /pollTurk every minute. Rather than get
unanswered questions, /pollTurk will get a list of recently answered
questions and get their responses, sms the responses, and mark the HIT
as "reviewing". This way is more efficient as it results in less hits to
amazon. This is important as the opperation needs to be completed in
less than a minute before it is started again.